byungmin : 회원가입, 유저챌린지 조회 리팩토링
