+++ build.gradle
@@ -38,6 +38,8 @@ dependencies {
 
     // validation
     implementation 'org.springframework.boot:spring-boot-starter-validation'
+    // aop
+    implementation 'org.springframework.boot:spring-boot-starter-aop'
     // sentry
     implementation 'io.sentry:sentry-spring-boot-starter:4.3.0'
